% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{calculate_distance}
\alias{calculate_distance}
\title{calculate_distance
Calculate distance between two strings}
\usage{
calculate_distance(x, y)
}
\arguments{
\item{x}{a string}

\item{y}{a string}
}
\value{
the distance between two strings
}
\description{
calculate_distance
Calculate distance between two strings
}
\examples{
calculate_distance("I am testing.","Testing is ")

}
