% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{clean_doi}
\alias{clean_doi}
\title{clean_doi}
\usage{
clean_doi(doi)
}
\arguments{
\item{doi}{a doi string}
}
\value{
the values of the node
}
\description{
clean_doi
}
\examples{
clean_doi("https://doi.org/10.1212/01.wnl.0000260060.60870.89")

}
