% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{clean_title}
\alias{clean_title}
\title{clean_title
clean a title, preparing other process}
\usage{
clean_title(x)
}
\arguments{
\item{x}{a string}

\item{y}{a string}
}
\value{
the distance between two strings
}
\description{
clean_title
clean a title, preparing other process
}
\examples{
clean_title("I&rsquo;m testing.")

}
