% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{compare_year}
\alias{compare_year}
\title{compare_year
compare two publication years is the same. Sometimes, the year has one year different for the same publication in two different databases}
\usage{
compare_year(year1, year2)
}
\arguments{
\item{year1}{a number}

\item{year2}{a number}
}
\value{
the distance between two strings
}
\description{
compare_year
compare two publication years is the same. Sometimes, the year has one year different for the same publication in two different databases
}
\examples{
compare_year(2001,2000)

}
