% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{DownloadJsonWithPmidsBatch}
\alias{DownloadJsonWithPmidsBatch}
\title{DownloadJsonWithPmidsBatch}
\usage{
DownloadJsonWithPmidsBatch(
  pmids,
  endpoint = "efetch",
  fileBaseName = "test.json"
)
}
\arguments{
\item{pmids}{a string of character. PubMed central Id}

\item{endpoint}{a string of characters. The API endpoint to use. e.g. "esummary", "efetch"}

\item{fileBaseName}{a string of character. The base name of the to be saved xml files}
}
\value{
the output file names
}
\description{
DownloadJsonWithPmidsBatch
}
\examples{
DownloadJsonWithPmidsBatch(c("28852052", "29041955"), endpoint = "efetch", fileBaseName="test.json")

}
