% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{extractAbstractFromPmcidEfetchDoc}
\alias{extractAbstractFromPmcidEfetchDoc}
\title{extractAbstractFromPmcidEfetchDoc}
\usage{
extractAbstractFromPmcidEfetchDoc(article)
}
\arguments{
\item{doc}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters. abstract
}
\description{
extractAbstractFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
extractAbstractFromPmcidEfetchDoc(doc)

}
