% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{extractAffliationFromPmcidEfetchDoc}
\alias{extractAffliationFromPmcidEfetchDoc}
\title{extractAffliationFromPmcidEfetchDoc}
\usage{
extractAffliationFromPmcidEfetchDoc(article, affIds = "all_affiliations")
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}

\item{affIds}{affiliation id to extract. Default is "all_affiliatins", which gives all the affiliations}
}
\value{
a string of characters.
}
\description{
extractAffliationFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id=c("2823164"), db="pmc", endpoint="efetch")
extractAffliationFromPmcidEfetchDoc(doc)

doc <- GetDoc(id="3339580", db="pmc", endpoint="efetch")
extractAffliationFromPmcidEfetchDoc(doc)

doc <- GetDoc(id="3324826", db="pmc", endpoint="efetch")
extractAffliationFromPmcidEfetchDoc(doc)

}
