% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{extractAuthorsFromPmcidEfetchDoc}
\alias{extractAuthorsFromPmcidEfetchDoc}
\title{extractAuthorsFromPmcidEfetchDoc}
\usage{
extractAuthorsFromPmcidEfetchDoc(article)
}
\arguments{
\item{doc}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
extractAuthorsFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="2823164", db="pmc", endpoint="efetch")
extractAuthorsFromPmcidEfetchDoc(doc)

}
