% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ExtractCorrespondindAuthorAffliationFromPmcidEfetchDoc}
\alias{ExtractCorrespondindAuthorAffliationFromPmcidEfetchDoc}
\title{ExtractCorrespondindAuthorAffliationFromPmcidEfetchDoc}
\usage{
ExtractCorrespondindAuthorAffliationFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
ExtractCorrespondindAuthorAffliationFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
ExtractCorrespondindAuthorAffliationFromPmcidEfetchDoc(doc)

doc <- GetDoc(id="4415024", db="pmc", endpoint="efetch")
ExtractCorrespondindAuthorAffliationFromPmcidEfetchDoc(doc)

}
