% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ExtractCorrespondingAuthorFromPmcidEfetchDoc}
\alias{ExtractCorrespondingAuthorFromPmcidEfetchDoc}
\title{ExtractCorrespondingAuthorFromPmcidEfetchDoc}
\usage{
ExtractCorrespondingAuthorFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
ExtractCorrespondingAuthorFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="4415024", db="pmc", endpoint="efetch")
ExtractCorrespondingAuthorFromPmcidEfetchDoc(doc)

doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
ExtractCorrespondingAuthorFromPmcidEfetchDoc(doc)

doc <- GetDoc(id="3339580", db="pmc", endpoint="efetch")
ExtractCorrespondingAuthorFromPmcidEfetchDoc(doc)
}
