% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{extractCorrespondingAuthorIdFromPmcidEfetchDoc}
\alias{extractCorrespondingAuthorIdFromPmcidEfetchDoc}
\title{extractCorrespondingAuthorIdFromPmcidEfetchDoc}
\usage{
extractCorrespondingAuthorIdFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
extractCorrespondingAuthorIdFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="3339580", db="pmc", endpoint="efetch")
extractCorrespondingAuthorIdFromPmcidEfetchDoc(doc)

doc <- GetDoc(id="3892617", db="pmc", endpoint="efetch")
extractCorrespondingAuthorIdFromPmcidEfetchDoc(doc)

doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
extractCorrespondingAuthorIdFromPmcidEfetchDoc(doc)

doc <- GetDoc(id="4415024", db="pmc", endpoint="efetch")
extractCorrespondingAuthorIdFromPmcidEfetchDoc(doc)

}
