% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ExtractEmailsFromPmcidEfetchDoc}
\alias{ExtractEmailsFromPmcidEfetchDoc}
\title{ExtractEmailsFromPmcidEfetchDoc}
\usage{
ExtractEmailsFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
ExtractEmailsFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="2823164", db="pmc", endpoint="efetch")
ExtractEmailsFromPmcidEfetchDoc(doc)

}
