% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ExtractFullTextFromPmcidEfetchDoc}
\alias{ExtractFullTextFromPmcidEfetchDoc}
\title{ExtractFullTextFromPmcidEfetchDoc}
\usage{
ExtractFullTextFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters. fulltext
}
\description{
ExtractFullTextFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
ExtractFullTextFromPmcidEfetchDoc(doc)

}
