% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ExtractFullTextToFile}
\alias{ExtractFullTextToFile}
\title{ExtractFullTextToFile}
\usage{
ExtractFullTextToFile(xmlFilePathFull, overwirte = F)
}
\arguments{
\item{xmlFilePathFull}{xml file full path}

\item{overwirte}{whether or not overwrite the text file if already exists}
}
\value{
textfilename, a string
}
\description{
ExtractFullTextToFile
}
