% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ExtractJournalFromPmcidEfetchDoc}
\alias{ExtractJournalFromPmcidEfetchDoc}
\title{ExtractJournalFromPmcidEfetchDoc}
\usage{
ExtractJournalFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
ExtractJournalFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="2823164", db="pmc", endpoint="efetch")
ExtractJournalFromPmcidEfetchDoc(doc)

}
