% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ExtractPmidFromPmcidEfetchDoc}
\alias{ExtractPmidFromPmcidEfetchDoc}
\title{ExtractPmidFromPmcidEfetchDoc}
\usage{
ExtractPmidFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
ExtractPmidFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="2823164", db="pmc", endpoint="efetch")
ExtractPmidFromPmcidEfetchDoc(doc)

}
