% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ExtractPublisherFromPmcidEfetchDoc}
\alias{ExtractPublisherFromPmcidEfetchDoc}
\title{ExtractPublisherFromPmcidEfetchDoc}
\usage{
ExtractPublisherFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
ExtractPublisherFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="2823164", db="pmc", endpoint="efetch")
ExtractPublisherFromPmcidEfetchDoc(doc)

}
