% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{extractStudyTypeFromPmcidEfetchDoc}
\alias{extractStudyTypeFromPmcidEfetchDoc}
\title{extractStudyTypeFromPmcidEfetchDoc}
\usage{
extractStudyTypeFromPmcidEfetchDoc(article)
}
\arguments{
\item{doc}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters. study type
}
\description{
extractStudyTypeFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
extractStudyTypeFromPmcidEfetchDoc(doc)

}
