% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{extractTitleFromPmcidEfetchDoc}
\alias{extractTitleFromPmcidEfetchDoc}
\title{extractTitleFromPmcidEfetchDoc}
\usage{
extractTitleFromPmcidEfetchDoc(article)
}
\arguments{
\item{doc}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters. title
}
\description{
extractTitleFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
extractTitleFromPmcidEfetchDoc(doc)

}
