% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetBaselink}
\alias{GetBaselink}
\title{GetBaselink}
\usage{
GetBaselink(db, ids, apiKey = "", email = "")
}
\arguments{
\item{db}{a string of characters, valid database name from NCBI: https://www.ncbi.nlm.nih.gov/books/NBK25497/#chapter2.chapter2_table1}

\item{ids}{a string of characters, valid id matching the database}

\item{apiKey}{a string of characters, API Key of the user}

\item{email}{a string of characters, email address of the user}
}
\value{
a list of links
}
\description{
GetBaselink
}
\examples{
links <- GetBaselink("pubmed", "4804230", "","")

}
