% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{GetContentByGetLink}
\alias{GetContentByGetLink}
\title{GetContentByGetLink}
\usage{
GetContentByGetLink(link, waitTime = 0.3)
}
\arguments{
\item{link}{a string of characters}

\item{waitTime}{a number. Waiting of the program}
}
\value{
a string of characters of the returned content
}
\description{
GetContentByGetLink
}
\examples{
baselink <- "https://eutils.ncbi.nlm.nih.gov/entrez/eutils/"
link <- paste0(baselink, "efetch.fcgi?db=pubmed&rettype=xml&id=26502666")
content <- GetContentByGetLink(link, 0.3)

}
