% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetContentWithLink}
\alias{GetContentWithLink}
\title{GetContentWithLink}
\usage{
GetContentWithLink(link)
}
\arguments{
\item{link}{a string of characters}
}
\value{
a string of characters of the returned content
}
\description{
GetContentWithLink
}
\examples{
content <- GetContentWithLink("https://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi?db=pubmed&rettype=xml&id=26502666")

}
