% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{GetDoc}
\alias{GetDoc}
\title{GetDoc}
\usage{
GetDoc(
  endpoint = "",
  db = "",
  id = "",
  apiKey = "",
  email = "",
  term = "",
  reldate = "",
  datetype = "",
  retmax = 1000,
  usehistory = "y",
  retstart = "",
  WebEnv = "",
  cmd = "",
  queryKey = ""
)
}
\arguments{
\item{endpoint}{a string of characters, valid endpoint of NCBI Utilities, e.g. "eserach", "efetch","elink", "esummary"}

\item{db}{a string of characters. Target database about which to gather statistics. Values must match https://www.ncbi.nlm.nih.gov/books/n/helpeutils/chapter2/#chapter2.chapter2_table1}

\item{id}{a list of strings. UID list. Either a single UID or a UID list}

\item{apiKey}{a string of characters. API Key of the user. NCBI will begin enforcing the practice of using an API key for sites that post more than 3 requests per second.}

\item{email}{a string of characters. E-mail address of the E-utility user.}

\item{term}{a string of characters. Entrez text query.}

\item{reldate}{a string of characters, When reldate is set to an integer n, the search returns only those items that have a date specified by datetype within the last n days.}

\item{datetype}{a string of characters. Type of date used to limit a search. The allowed values vary between Entrez databases, but common values are 'mdat' (modification date), 'pdat' (publication date) and 'edat' (Entrez date).}

\item{retmax}{a string of characters. Total number of UIDs from the retrieved set to be shown in the XML output (default=20). By default, ESearch only includes the first 20 UIDs retrieved in the XML output.}

\item{usehistory}{a string of characters. When usehistory is set to 'y', ESearch will post the UIDs resulting from the search operation onto the History server so that they can be used directly in a subsequent E-utility call. Also, usehistory must be set to 'y' for ESearch to interpret query key values included in term or to accept a WebEnv as input.}

\item{retstart}{a string of characters.Sequential index of the first UID in the retrieved set to be shown in the XML output. This parameter can be used in conjunction with retmax to download an arbitrary subset of UIDs retrieved from a search.}

\item{WebEnv}{a string of characters. Web environment string returned from a previous ESearch, EPost or ELink call. When provided, ESearch will post the results of the search operation to this pre-existing}

\item{cmd}{a string of characters. ELink command mode. The command mode specified which function ELink will perform. Some optional parameters only function for certain values of &cmd}

\item{queryKey}{a string of characters. an integer label called a query key}
}
\value{
a XMLInternalDocument
}
\description{
GetDoc
}
\examples{
GetDoc(db = "pubmed", endpoint = "esearch", term="stroke", retmax=5)
GetDoc(id = "5575286", db = "pmc", endpoint = "efetch")

}
