% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{GetMetaDataFromPmcidBatch}
\alias{GetMetaDataFromPmcidBatch}
\title{GetMetaDataFromPmcidBatch}
\usage{
GetMetaDataFromPmcidBatch(pmcids, apiKey = "", email = "", writeFileName = "")
}
\arguments{
\item{pmcids}{a string of character. PubMed central Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{writeFileName}{a string of character. The base name of the to be saved xml files}
}
\value{
a nx3 data frame. With three columns: pmcid, pmid, doi
}
\description{
GetMetaDataFromPmcidBatch
}
\examples{
GetMetaDataFromPmcidBatch(c("5575286", "4804230"))

}
