% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{GetPmidDoiFromPmcid}
\alias{GetPmidDoiFromPmcid}
\title{GetPmidDoiFromPmcid}
\usage{
GetPmidDoiFromPmcid(pmcid, writeFileName = "")
}
\arguments{
\item{pmcid}{a string of character. PubMed central Id}

\item{writeFileName}{a string of characters.}
}
\value{
a string: pmid
}
\description{
GetPmidDoiFromPmcid
}
\examples{
GetPmidDoiFromPmcid(c("5575286", "4804230"))

}
