% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetPmidsWithSearch}
\alias{GetPmidsWithSearch}
\title{GetPmidsWithSearch}
\usage{
GetPmidsWithSearch(searchTerm)
}
\arguments{
\item{searchTerm}{a string of characters. search string}
}
\value{
the output file names
}
\description{
GetPmidsWithSearch
}
\examples{
searchTerm = "pinkeye"
GetPmidsWithSearch(searchTerm)

}
