% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetSearchSummaryWithSearch}
\alias{GetSearchSummaryWithSearch}
\title{GetSearchSummaryWithSearch}
\usage{
GetSearchSummaryWithSearch(searchTerm)
}
\arguments{
\item{searchTerm}{a string of characters. search string}
}
\value{
a list of strings - searchSummary
}
\description{
GetSearchSummaryWithSearch
}
\examples{
 GetSearchSummaryWithSearch("pinkeye")

}
