% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetUrlsFromPmid}
\alias{GetUrlsFromPmid}
\title{GetUrlsFromPmid}
\usage{
GetUrlsFromPmid(pmid, apiKey, fulltext = T)
}
\arguments{
\item{pmid}{a number or a string of characters. The number of pmid.}

\item{fulltext}{a boolean}
}
\value{
a list of characters. A list of urls. Return fulltext urls if fulltext parameter is T.
Return NULL if none is found.
}
\description{
Get related full text urls from given pmid
}
\examples{
 pmid <- "28852052"
apiKey <- ""
url <-  GetUrlsFromPubMed(pmid, apiKey)
print(url)

}
