% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetXmlDocFromIds}
\alias{GetXmlDocFromIds}
\title{GetXmlDocFromIds}
\usage{
GetXmlDocFromIds(ids, database, target, apiKey = "", email = "",
  waitTime = "")
}
\arguments{
\item{ids}{a string of chacters. Valid NCBI ids}

\item{database}{a string of chacters. Valid NCBI}

\item{target}{a string of chacters. "efetch", "esummary", "elink"}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{waitTime}{a number. Time to wait during the http quest}
}
\value{
a XMLInternalDocument
}
\description{
GetXmlDocFromIds
}
\examples{
GetXmlDocFromIds("5575286", "pmc", "efetch", "", "", 0.3)

}
