% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ReadMetaDataFromPMCnxml}
\alias{ReadMetaDataFromPMCnxml}
\title{ReadMetaDataFromPMCnxml}
\usage{
ReadMetaDataFromPMCnxml(
  link,
  columns = c("pmid", "title", "abstract", "fulltext", "studytype", "journal",
    "publisher", "publicationDate", "authors", "emails", "affiliations",
    "correspondingAuthors", "correspondingAuthorAffs")
)
}
\arguments{
\item{link}{a XMLInternalDocument, a XMLAbstractDocument}

\item{columns}{list of output columns}
}
\value{
a 1x3 data frame. With three columns: pmcid, pmid, doi
}
\description{
ReadMetaDataFromPMCnxml
}
