% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{Retrive_pmid_with_TitleYear}
\alias{Retrive_pmid_with_TitleYear}
\title{Retrive_pmid_with_TitleYear
Year is optional}
\usage{
Retrive_pmid_with_TitleYear(title, year = NULL)
}
\arguments{
\item{title}{a string of characters. title of the publication}

\item{year}{a integer. Publication year. Optional}
}
\value{
a list of strings - searchSummary
}
\description{
Retrive_pmid_with_TitleYear
Year is optional
}
\examples{
 
title <- "Safety and efficacy of daclizumab in relapsing-remitting multiple sclerosis: 3-year results from the SELECTED open-label extension study."
year <- 2016 
Retrive_pmid_with_TitleYear(title, year)


}
