% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveAbstractFromPubmedEfetch}
\alias{RetriveAbstractFromPubmedEfetch}
\title{RetriveAbstractFromPubmedEfetch}
\usage{
RetriveAbstractFromPubmedEfetch(doc)
}
\arguments{
\item{doc}{an XMLInternalDocument class file read from a pubmed efetch xml file}
}
\value{
a string
}
\description{
RetriveAbstractFromPubmedEfetch
}
\examples{
 doc <- GetDoc(id = "28852052", db = "pubmed", endpoint = "efetch")
 RetriveAbstractFromPubmedEfetch(doc)
}
