% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveDOIFromPubmedEfetch}
\alias{RetriveDOIFromPubmedEfetch}
\title{RetriveDOIFromPubmedEfetch}
\usage{
RetriveDOIFromPubmedEfetch(doc)
}
\arguments{
\item{doc}{an XMLInternalDocument class file read from a pubmed efetch xml file}
}
\value{
a string
}
\description{
RetriveDOIFromPubmedEfetch
}
\examples{
doc <- GetDoc(id = "9214527", db = "pubmed", endpoint = "efetch")
RetriveDOIFromPubmedEfetch(doc)

}
