% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveIBSNFromPubmedEfetch}
\alias{RetriveIBSNFromPubmedEfetch}
\title{RetriveIBSNFromPubmedEfetch}
\usage{
RetriveIBSNFromPubmedEfetch(doc)
}
\arguments{
\item{doc}{an XMLInternalDocument class file read from a pubmed efetch xml file}
}
\value{
a string
}
\description{
RetriveIBSNFromPubmedEfetch
}
\examples{
 doc <- GetDoc(id = "31524133", db = "pubmed", endpoint = "efetch")
RetriveIBSNFromPubmedEfetch(doc)

}
