% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveJournalFromPmidEfetcXML}
\alias{RetriveJournalFromPmidEfetcXML}
\title{RetriveJournalFromPmidEfetcXML}
\usage{
RetriveJournalFromPmidEfetcXML(pmid, apiKey = "", email = "")
}
\arguments{
\item{pmid}{a string of character. PubMed Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
RetriveJournalFromPmidEfetcXML
}
\examples{
 journal <- RetriveJournalFromPmidEfetcXML(c("28852052", "29041955","31230181"))
}
