% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveJournalFromPubmedEfetch}
\alias{RetriveJournalFromPubmedEfetch}
\title{RetriveJournalFromPubmedEfetch}
\usage{
RetriveJournalFromPubmedEfetch(doc)
}
\arguments{
\item{doc}{an XMLInternalDocument class file read from a pubmed efetch xml file}
}
\value{
a string
}
\description{
RetriveJournalFromPubmedEfetch
}
\examples{
 doc <- GetDoc(id = "28852052", db = "pubmed", endpoint = "efetch")
 RetriveJournalFromPubmedEfetch(doc)
}
