% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveJournalWithPmids}
\alias{RetriveJournalWithPmids}
\title{RetriveJournalWithPmids}
\usage{
RetriveJournalWithPmids(pmid)
}
\arguments{
\item{pmid}{a string of character. PubMed Id}
}
\value{
a nx3 data frame
}
\description{
RetriveJournalWithPmids
}
\examples{
 journal <- RetriveJournalWithPmids(c("28852052", "29041955","31230181"))
}
