% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveKeywordsFromPubmedEfetch}
\alias{RetriveKeywordsFromPubmedEfetch}
\title{RetriveKeywordsFromPubmedEfetch}
\usage{
RetriveKeywordsFromPubmedEfetch(doc)
}
\arguments{
\item{doc}{an XMLInternalDocument class file read from a pubmed efetch xml file}
}
\value{
a string
}
\description{
RetriveKeywordsFromPubmedEfetch
}
\examples{
 doc <- GetDoc(id = "31524133", db = "pubmed", endpoint = "efetch")
RetriveKeywordsFromPubmedEfetch(doc)

}
