% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveMetaDataFromPmids}
\alias{RetriveMetaDataFromPmids}
\title{RetriveMetaDataFromPmids}
\usage{
RetriveMetaDataFromPmids(pmids, apiKey)
}
\arguments{
\item{pmids}{a string of character. PubMed Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}
}
\description{
RetriveMetaDataFromPmids
}
\examples{
 pmids <- c("28852052", "29041955")
apiKey <- ""
metaData <-  RetriveMetaDataFromPmids(pmids, apiKey)
print(metaData)

}
