% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveMetaDataFromPubmedEfetch}
\alias{RetriveMetaDataFromPubmedEfetch}
\title{RetriveMetaDataFromPubmedEfetch}
\usage{
RetriveMetaDataFromPubmedEfetch(doc)
}
\arguments{
\item{doc}{an XMLInternalDocument class file read from a pubmed efetch xml file}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
RetriveMetaDataFromPubmedEfetch
}
\examples{
 doc <- GetDoc(id = c("28852052", "29041955","31230181"), db = "pubmed", endpoint = "efetch")
metaData <- RetriveMetaDataFromPubmedEfetch(doc)
}
