% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveMetaDataFromSearch}
\alias{RetriveMetaDataFromSearch}
\title{RetriveMetaDataFromSearch}
\usage{
RetriveMetaDataFromSearch(searchTerm, columns = "", outputFilename = "")
}
\arguments{
\item{searchTerm}{a string of character. search term}

\item{columns}{a list of string of requested column names: "pmid","pmcid", "journal", "journalCountry","publicationYear","funders","authors","affiliations","title","abstract","isbn","volume","issue","pages", "keywords","doi"}

\item{outputFilename}{a string of characters. Output XML file name}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
RetriveMetaDataFromSearch
}
\examples{
searchTerm <-  "pinkeye"
metaData <- RetriveMetaDataFromSearch(searchTerm)

}
