% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetrivePagesFromPubmedEfetch}
\alias{RetrivePagesFromPubmedEfetch}
\title{RetrivePagesFromPubmedEfetch}
\usage{
RetrivePagesFromPubmedEfetch(doc)
}
\arguments{
\item{doc}{an XMLInternalDocument class file read from a pubmed efetch xml file}
}
\value{
a string
}
\description{
RetrivePagesFromPubmedEfetch
}
\examples{
 doc <- GetDoc(id = "31524133", db = "pubmed", endpoint = "efetch")
RetrivePagesFromPubmedEfetch(doc)

}
