% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetrivePmcidWithPmids}
\alias{RetrivePmcidWithPmids}
\title{RetrivePmcidWithPmids}
\usage{
RetrivePmcidWithPmids(pmid)
}
\arguments{
\item{pmid}{a string of character. PubMed Id}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
RetrivePmcidWithPmids
}
\examples{
 pmcids <- RetrivePmcidWithPmids(c("28852052", "29041955","31230181"))
}
