% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveUrlfromELinkXML}
\alias{RetriveUrlfromELinkXML}
\title{RetriveUrlfromELinkXML}
\usage{
RetriveUrlfromELinkXML(doc, category = "All")
}
\arguments{
\item{doc}{a XMLInternalDocument.}

\item{category}{a string of characters. A link category. Defaul is "All"}
}
\value{
a list of characters. A list of urls. Return fulltext urls if fulltext parameter is T.
Return NULL if none is found.
}
\description{
Retrive Url from Elink xml doc
}
\examples{
doc <- GetDoc(id="28852052", db="pubmed", endpoint="elink", cmd = "llinks")
RetriveUrlfromELinkXML(doc, "Full Text Sources")

}
