% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveUrlsFromPmids}
\alias{RetriveUrlsFromPmids}
\title{RetriveUrlsFromPmids}
\usage{
RetriveUrlsFromPmids(pmids, apiKey = "", email = "", waitTime = 0.3,
  fulltext = TRUE)
}
\arguments{
\item{pmids}{a list of numbers or characters. The number of pmid.}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{waitTime}{a number. Waiting of the program}

\item{fulltext}{a boolean. If TRUE, function only searches for full text link}
}
\value{
a list of characters. A list of urls. Return fulltext urls if fulltext parameter is T.
Return NULL if none is found.
}
\description{
Retrive urls from pmids
}
\examples{
pmids <- c("28852052", "29041955")
urls <-  RetriveUrlsFromPmids(pmids, "", "", 0.3, TRUE)
print(urls)

}
