% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveUrlsFromPmidsBatch}
\alias{RetriveUrlsFromPmidsBatch}
\title{RetriveUrlsFromPmidsBatch}
\usage{
RetriveUrlsFromPmidsBatch(pmids, fulltext = TRUE)
}
\arguments{
\item{pmids}{a string of character. PubMed central Id}

\item{fulltext}{a boolean. Whether or not only retrive full text}
}
\value{
a nx3 data frame. With three columns: pmcid, pmid, doi
}
\description{
this function is good for retrive data directly
}
\examples{
RetriveUrlsFromPmidsBatch(c("28852052", "29041955"))

}
