% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveUrlsFromPmidsParallel}
\alias{RetriveUrlsFromPmidsParallel}
\title{RetriveUrlsFromPmidsParallel}
\usage{
RetriveUrlsFromPmidsParallel(pmids, apiKey, fulltext = T)
}
\arguments{
\item{pmids}{a list of numbers or characters. The number of pmid.}

\item{fulltext}{a boolean}
}
\value{
a list of characters. A list of urls. Return fulltext urls if fulltext parameter is T.
Return NULL if none is found.
}
\description{
Retrive urls from pmids
}
\examples{
pmids <- c("28852052", "29041955")
apiKey <- ""
urls <-  RetriveUrlsFromPmidsParallel(pmids, apiKey)
print(urls)

}
