% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveXmlNodeValuefromContent}
\alias{RetriveXmlNodeValuefromContent}
\title{RetriveXmlNodeValuefromContent}
\usage{
RetriveXmlNodeValuefromContent(content, nodePosition)
}
\arguments{
\item{content}{a string of characters, content received from GET request}

\item{nodePosition}{a string of characters, valid nodePosition of the XML}
}
\value{
the values of the request node
}
\description{
RetriveXmlNodeValuefromContent
}
