% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveXmlNodeValuefromDoc}
\alias{RetriveXmlNodeValuefromDoc}
\title{RetriveXmlNodeValuefromDoc}
\usage{
RetriveXmlNodeValuefromDoc(doc, nodePosition)
}
\arguments{
\item{doc}{the parsed XML file}

\item{nodePosition}{the node position of the xml file}
}
\value{
the values of the node
}
\description{
RetriveXmlNodeValuefromDoc
}
