% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{DownloadMetaDataWithPmcidsBatch}
\alias{DownloadMetaDataWithPmcidsBatch}
\title{DownloadMetaDataWithPmcidsBatch}
\usage{
DownloadMetaDataWithPmcidsBatch(
  pmcids,
  apiKey = "",
  email = "",
  fileBaseName = ""
)
}
\arguments{
\item{pmcids}{a string of character. PubMed central Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{fileBaseName}{a string of character. The base name of the to be saved xml files}
}
\value{
a nx3 data frame. With three columns: pmcid, pmid, doi
}
\description{
DownloadMetaDataWithPmcidsBatch
}
\examples{
DownloadMetaDataWithPmcidsBatch(c("5575286", "4804230"), fileBaseName="test.xml")

}
