% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{DownloadMetaDataWithPmcidsBatch}
\alias{DownloadMetaDataWithPmcidsBatch}
\title{DownloadMetaDataWithPmcidsBatch}
\usage{
DownloadMetaDataWithPmcidsBatch(pmcids, apiKey, email, waitTime,
  fileBaseName)
}
\arguments{
\item{pmcids}{a string of character. PubMed central Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{waitTime}{a number. Waiting of the program}

\item{fileBaseName}{a string of character. The base name of the to be saved xml files}
}
\value{
a nx3 data frame. With three columns: pmcid, pmid, doi
}
\description{
DownloadMetaDataWithPmcidsBatch
}
\examples{
DownloadMetaDataWithPmcidsBatch(c("5575286", "4804230"), "",  "", 0, "test.xml")

}
