% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{DownloadXMLWithPmidsBatch}
\alias{DownloadXMLWithPmidsBatch}
\title{DownloadMetaDataWithPmidsBatch}
\usage{
DownloadXMLWithPmidsBatch(pmids, apiKey = "", endpoint = "esummary",
  email = "", fileBaseName = "")
}
\arguments{
\item{pmids}{a string of character. PubMed central Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{endpoint}{a string of characters. The API endpoint to use. e.g. "esummary", "efetch"}

\item{email}{a string of characters. Your email address}

\item{fileBaseName}{a string of character. The base name of the to be saved xml files}
}
\value{
a nx3 data frame. With three columns: pmcid, pmid, doi
}
\description{
DownloadMetaDataWithPmidsBatch
}
\examples{
DownloadXMLWithPmidsBatch(c("28852052", "29041955"), endpoint = "esummary", fileBaseName="test.xml")

}
