% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ExtractAbstractFromPmcidEfetchDoc}
\alias{ExtractAbstractFromPmcidEfetchDoc}
\title{ExtractAbstractFromPmcidEfetchDoc}
\usage{
ExtractAbstractFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters. abstract
}
\description{
ExtractAbstractFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
ExtractAbstractFromPmcidEfetchDoc(doc)

}
