% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{extractCorrespondindAuthorAffliationFromPmcidEfetchDoc}
\alias{extractCorrespondindAuthorAffliationFromPmcidEfetchDoc}
\title{extractCorrespondindAuthorAffliationFromPmcidEfetchDoc}
\usage{
extractCorrespondindAuthorAffliationFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
extractCorrespondindAuthorAffliationFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
extractCorrespondindAuthorAffliationFromPmcidEfetchDoc(doc)

doc <- GetDoc(id="4415024", db="pmc", endpoint="efetch")
extractCorrespondindAuthorAffliationFromPmcidEfetchDoc(doc)

}
