% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{extractCorrespondingAuthorFromPmcidEfetchDoc}
\alias{extractCorrespondingAuthorFromPmcidEfetchDoc}
\title{extractCorrespondingAuthorFromPmcidEfetchDoc}
\usage{
extractCorrespondingAuthorFromPmcidEfetchDoc(article)
}
\arguments{
\item{doc}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
extractCorrespondingAuthorFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="3324826", db="pmc", endpoint="efetch")
extractCorrespondingAuthorFromPmcidEfetchDoc(doc)

doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
extractCorrespondingAuthorFromPmcidEfetchDoc(doc)

doc <- GetDoc(id=c("2823164"), db="pmc", endpoint="efetch")
extractCorrespondingAuthorFromPmcidEfetchDoc(doc)
}
