% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ExtractCorrespondingAuthorIdFromPmcidEfetchDoc}
\alias{ExtractCorrespondingAuthorIdFromPmcidEfetchDoc}
\title{ExtractCorrespondingAuthorIdFromPmcidEfetchDoc}
\usage{
ExtractCorrespondingAuthorIdFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
ExtractCorrespondingAuthorIdFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="3339580", db="pmc", endpoint="efetch")
ExtractCorrespondingAuthorIdFromPmcidEfetchDoc(doc)

doc <- GetDoc(id="3892617", db="pmc", endpoint="efetch")
ExtractCorrespondingAuthorIdFromPmcidEfetchDoc(doc)

doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
ExtractCorrespondingAuthorIdFromPmcidEfetchDoc(doc)

doc <- GetDoc(id="4415024", db="pmc", endpoint="efetch")
ExtractCorrespondingAuthorIdFromPmcidEfetchDoc(doc)

}
