% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{extractEmailsFromPmcidEfetchDoc}
\alias{extractEmailsFromPmcidEfetchDoc}
\title{extractEmailsFromPmcidEfetchDoc}
\usage{
extractEmailsFromPmcidEfetchDoc(article)
}
\arguments{
\item{doc}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
extractEmailsFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="2823164", db="pmc", endpoint="efetch")
extractEmailsFromPmcidEfetchDoc(doc)

}
