% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ExtractEpubDateFromPmcidEfetchDoc}
\alias{ExtractEpubDateFromPmcidEfetchDoc}
\title{ExtractEpubDateFromPmcidEfetchDoc}
\usage{
ExtractEpubDateFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string: publicationDate
}
\description{
ExtractEpubDateFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="5304250",db= "pmc", endpoint="efetch")
ExtractEpubDateFromPmcidEfetchDoc(doc)
doc <- GetDoc(id="4415024",db= "pmc", endpoint="efetch")
ExtractEpubDateFromPmcidEfetchDoc(doc)
doc <- GetDoc(id="4804230",db= "pmc", endpoint="efetch")
ExtractEpubDateFromPmcidEfetchDoc(doc)

}
