% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{extractFullTextFromPmcidEfetchDoc}
\alias{extractFullTextFromPmcidEfetchDoc}
\title{extractFullTextFromPmcidEfetchDoc}
\usage{
extractFullTextFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters. fulltext
}
\description{
extractFullTextFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
extractFullTextFromPmcidEfetchDoc(doc)

}
