% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{extractJournalFromPmcidEfetchDoc}
\alias{extractJournalFromPmcidEfetchDoc}
\title{extractJournalFromPmcidEfetchDoc}
\usage{
extractJournalFromPmcidEfetchDoc(article)
}
\arguments{
\item{doc}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
extractJournalFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="2823164", db="pmc", endpoint="efetch")
extractJournalFromPmcidEfetchDoc(doc)

}
