% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{extractPmidFromPmcidEfetchDoc}
\alias{extractPmidFromPmcidEfetchDoc}
\title{extractPmidFromPmcidEfetchDoc}
\usage{
extractPmidFromPmcidEfetchDoc(article)
}
\arguments{
\item{doc}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
extractPmidFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="2823164", db="pmc", endpoint="efetch")
extractPmidFromPmcidEfetchDoc(doc)

}
