% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{extractPublisherFromPmcidEfetchDoc}
\alias{extractPublisherFromPmcidEfetchDoc}
\title{extractPublisherFromPmcidEfetchDoc}
\usage{
extractPublisherFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters.
}
\description{
extractPublisherFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="2823164", db="pmc", endpoint="efetch")
extractPublisherFromPmcidEfetchDoc(doc)

}
