% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ExtractStudyTypeFromPmcidEfetchDoc}
\alias{ExtractStudyTypeFromPmcidEfetchDoc}
\title{ExtractStudyTypeFromPmcidEfetchDoc}
\usage{
ExtractStudyTypeFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters. study type
}
\description{
ExtractStudyTypeFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
ExtractStudyTypeFromPmcidEfetchDoc(doc)

}
