% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{ExtractTitleFromPmcidEfetchDoc}
\alias{ExtractTitleFromPmcidEfetchDoc}
\title{ExtractTitleFromPmcidEfetchDoc}
\usage{
ExtractTitleFromPmcidEfetchDoc(article)
}
\arguments{
\item{article}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a string of characters. title
}
\description{
ExtractTitleFromPmcidEfetchDoc
}
\examples{
doc <- GetDoc(id="4405051", db="pmc", endpoint="efetch")
ExtractTitleFromPmcidEfetchDoc(doc)

}
