% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetMetaDataFromPmcid}
\alias{GetMetaDataFromPmcid}
\title{GetMetaDataFromPmcid}
\usage{
GetMetaDataFromPmcid(pmcid, apiKey = "", email = "", waitTime = 0.3,
  writeFileName = "")
}
\arguments{
\item{pmcid}{a string of character. PubMed Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{waitTime}{a number. Waiting of the program}

\item{writeFileName}{a string of characters. The file name you would like to download to.}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
GetMetaDataFromPmcid
}
\examples{
GetMetaDataFromPmcid(c("5575286", "4804230"))

}
