% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{GetMetaDataFromPmcid}
\alias{GetMetaDataFromPmcid}
\title{GetMetaDataFromPmcid}
\usage{
GetMetaDataFromPmcid(pmcid, apiKey = "", email = "", writeFileName = "")
}
\arguments{
\item{pmcid}{a string of character. PubMed Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{writeFileName}{a string of characters. The file name you would like to download to.}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
The number of pmcids sent should be less than 500.
}
\examples{
mydata <- GetMetaDataFromPmcid(c("5575286", "4804230"))

}
