% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetMetaDataFromPmid}
\alias{GetMetaDataFromPmid}
\title{GetMetaDataFromPmid}
\usage{
GetMetaDataFromPmid(pmid, apiKey)
}
\arguments{
\item{pmid}{a string of character. PubMed Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
GetMetaDataFromPmid
}
\examples{
pmid <- "28852052"
apiKey <- ""
metaData <-  GetMetaDataFromPmid(pmid, apiKey)
print(metaData)

}
