% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetMetaDataFromPmid}
\alias{GetMetaDataFromPmid}
\title{GetMetaDataFromPmid}
\usage{
GetMetaDataFromPmid(pmid, apiKey = "", email = "", waitTime = 0.3)
}
\arguments{
\item{pmid}{a string of character. PubMed Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{waitTime}{a number. Waiting of the program}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
GetMetaDataFromPmid
}
\examples{
pmid <- "28852052"
metaData <-  GetMetaDataFromPmid(pmid)
print(metaData)

}
