% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetPmidDoiFromPmcid}
\alias{GetPmidDoiFromPmcid}
\title{GetPmidDoiFromPmcid}
\usage{
GetPmidDoiFromPmcid(pmcid, apiKey, email)
}
\arguments{
\item{pmcid}{a string of character. PubMed central Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}
}
\value{
a string: pmid
}
\description{
GetPmidDoiFromPmcid
}
\examples{
pmcid <- "28852052"
apiKey <- ""
email <- ""
pmid <-  GetPmidDoiFromPmcid(pmcid, apiKey, email)
print(pmid)

}
