% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetPmidDoiFromPmcid}
\alias{GetPmidDoiFromPmcid}
\title{GetPmidDoiFromPmcid}
\usage{
GetPmidDoiFromPmcid(pmcid, apiKey = "", email = "", waitTime = 0.3,
  writeFileName = "")
}
\arguments{
\item{pmcid}{a string of character. PubMed central Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{waitTime}{a number. Waiting of the program}

\item{writeFileName}{a string of characters.}
}
\value{
a string: pmid
}
\description{
GetPmidDoiFromPmcid
}
\examples{
GetPmidDoiFromPmcid(c("5575286", "4804230"))

}
