% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetPmidDoiFromPmcidBatch}
\alias{GetPmidDoiFromPmcidBatch}
\title{GetPmidDoiFromPmcidBatch}
\usage{
GetPmidDoiFromPmcidBatch(pmcids, apiKey, email, waitTime)
}
\arguments{
\item{pmcids}{a string of character. PubMed central Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{waitTime}{a number. Waiting of the program}
}
\value{
a nx3 data frame. With three columns: pmcid, pmid, doi
}
\description{
GetPmidDoiFromPmcidBatch
}
\examples{
pmid <- GetPmidDoiFromPmcidBatch(c("5575286", "4804230"), "",  "", 0)

}
