% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcRead.R
\name{GetPmidDoiFromPmcidBatch}
\alias{GetPmidDoiFromPmcidBatch}
\title{GetPmidDoiFromPmcidBatch}
\usage{
GetPmidDoiFromPmcidBatch(pmcids)
}
\arguments{
\item{pmcids}{a string of character. PubMed central Id}
}
\value{
a nx3 data frame. With three columns: pmcid, pmid, doi
}
\description{
GetPmidDoiFromPmcidBatch
}
\examples{
GetPmidDoiFromPmcidBatch(c("5575286", "4804230"))


}
