% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetPmidsWithSearch}
\alias{GetPmidsWithSearch}
\title{GetPmidsWithSearch}
\usage{
GetPmidsWithSearch(searchTerm, apiKey = "", email = "")
}
\arguments{
\item{searchTerm}{a string of characters. search string}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}
}
\value{
the output file names
}
\description{
GetPmidsWithSearch
}
\examples{
GetPmidsWithSearch("pinkeye")

}
