% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetPmidsWithSearch}
\alias{GetPmidsWithSearch}
\title{GetPmidsWithSearch}
\usage{
GetPmidsWithSearch(searchTerm, apiKey = "")
}
\arguments{
\item{searchTerm}{a string of characters. search string}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}
}
\value{
the output file names
}
\description{
GetPmidsWithSearch
}
\examples{
term = "pinkeye"
GetPmidsWithSearch(term)

}
