% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{GetUrlsFromPmid}
\alias{GetUrlsFromPmid}
\title{GetUrlsFromPmid}
\usage{
GetUrlsFromPmid(pmid, apiKey = "", email = "", waitTime = 0.3,
  fulltext = TRUE)
}
\arguments{
\item{pmid}{a number or a string of characters. The number of pmid.}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{waitTime}{a number. Waiting of the program}

\item{fulltext}{a boolean. If TRUE, function only searches for full text link}
}
\value{
a list of characters. A list of urls. Return fulltext urls if fulltext parameter is T.
Return NULL if none is found.
}
\description{
Get related full text urls from given pmid
}
\examples{
 pmid <- "28852052"
url <-  GetUrlsFromPmid(pmid, "", "",0.3)

}
