% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{ReadMetaDataFromPmcidEfetchDoc}
\alias{ReadMetaDataFromPmcidEfetchDoc}
\title{ReadMetaDataFromPmcidEfetchDoc}
\usage{
ReadMetaDataFromPmcidEfetchDoc(doc)
}
\arguments{
\item{doc}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
a nx3 data frame. With three columns: pmcid, pmid, doi
}
\description{
ReadMetaDataFromPmcidEfetchDoc
}
\examples{
 doc <- GetXmlDocFromIds(c("2823164", "3324826", "3339580"), "pmc", "efetch", "", "", 0)
ReadMetaDataFromPmcidEfetchDoc(doc)

doc <- GetXmlDocFromIds(c("5575286","4405051","4804230"), "pmc", "efetch", "", "", 0)
ReadMetaDataFromPmcidEfetchDoc(doc)

}
