% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{ReadPmidDoiFromPmcidEsummaryDoc}
\alias{ReadPmidDoiFromPmcidEsummaryDoc}
\title{ReadPmidDoiFromPmcidEsummaryDoc}
\usage{
ReadPmidDoiFromPmcidEsummaryDoc(doc)
}
\arguments{
\item{doc}{a XMLInternalDocument, a XMLAbstractDocument}
}
\value{
extract pmid and doi for all articles in the doc
}
\description{
ReadPmidDoiFromPmcidEsummaryDoc
}
\examples{
 doc <- GetXmlDocFromIds(c("5575286", "4804230"), "pmc", "esummary", "", "", 0)
ReadPmidDoiFromPmcidEsummaryDoc(doc)

}
