% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{Retrive_pmid_with_TitleYear}
\alias{Retrive_pmid_with_TitleYear}
\title{Retrive_pmid_with_TitleYear
Year is optional.}
\usage{
Retrive_pmid_with_TitleYear(title, year = NULL, retmax = 1)
}
\arguments{
\item{title}{a string of characters. title of the publication}

\item{year}{a integer. Publication year. Optional}

\item{retmax}{a integer as the retmax parameter, the maximum id to render}
}
\value{
a list of strings - searchSummary
}
\description{
Retrive_pmid_with_TitleYear
Year is optional.
}
\examples{
title <- "Causes of Acute Stroke: A Patterned Approach."
Retrive_pmid_with_TitleYear(title)
Retrive_pmid_with_TitleYear(title, 2019)


}
