% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveFunderWithPmids}
\alias{RetriveFunderWithPmids}
\title{RetriveFunderWithPmids}
\usage{
RetriveFunderWithPmids(pmid, apiKey = "")
}
\arguments{
\item{pmid}{a string of character. PubMed Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
RetriveFunderWithPmids
}
\examples{
 RetriveFunderWithPmids(c("29041955","31230181"))
}
