% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveMetaDataFromPmids}
\alias{RetriveMetaDataFromPmids}
\title{RetriveMetaDataFromPmids}
\usage{
RetriveMetaDataFromPmids(pmid, apiKey = "", email = "",
  outputFilename = "")
}
\arguments{
\item{pmid}{a string of character. PubMed Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{outputFilename}{a string of characters. Output XML file name}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
RetriveMetaDataFromPmids
}
\examples{
 metaData <- RetriveMetaDataFromPmids(c("28852052", "29041955","31230181"), apiKey = "2ab938c0c3cfbfe6446b544c8e37c5a4e609")
}
