% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveMetaDataFromPmids}
\alias{RetriveMetaDataFromPmids}
\title{RetriveMetaDataFromPmids}
\usage{
RetriveMetaDataFromPmids(pmids, apiKey = "", email = "",
  waitTime = 0.3)
}
\arguments{
\item{pmids}{a string of character. PubMed Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{waitTime}{a number. Waiting of the program}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
RetriveMetaDataFromPmids
}
\examples{
 pmids <- c("28852052", "29041955")
metaData <-  RetriveMetaDataFromPmids(pmids)
print(metaData)


}
