% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveMetaDataFromPmids}
\alias{RetriveMetaDataFromPmids}
\title{RetriveMetaDataFromPmids}
\usage{
RetriveMetaDataFromPmids(pmid, outputFilename = "", columns = "")
}
\arguments{
\item{pmid}{a string of character. PubMed Id}

\item{outputFilename}{a string of characters. Output XML file name}

\item{columns}{the columns of output requested}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
RetriveMetaDataFromPmids
}
\examples{
 metaData <- RetriveMetaDataFromPmids(c("28852052", "29041955","31230181", "20704052"))
 metaData <- RetriveMetaDataFromPmids(c("20704052"))
}
