% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveMetaDataFromPmids}
\alias{RetriveMetaDataFromPmids}
\title{RetriveMetaDataFromPmids}
\usage{
RetriveMetaDataFromPmids(pmids, apiKey, email)
}
\arguments{
\item{pmids}{a string of character. PubMed Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}
}
\description{
RetriveMetaDataFromPmids
}
