% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveMetaDataFromPmidsBatch}
\alias{RetriveMetaDataFromPmidsBatch}
\title{RetriveMetaDataFromPmidsBatch}
\usage{
RetriveMetaDataFromPmidsBatch(
  pmids,
  outputFileBaseName = "",
  columns = "",
  grid = 500
)
}
\arguments{
\item{pmids}{a string of character. PubMed central Id}

\item{outputFileBaseName}{a string of characters. The base name of output xml files. If default, there will be no xml saved.}

\item{columns}{a list of string of requested column names: "pmid","pmcid", "journal", "journalCountry","publicationYear","funders","authors","affiliations","title","abstract","isbn","volume","issue","pages", "keywords","doi"}

\item{grid}{the number of ids to run in each loop}
}
\value{
a nx7 data frame. With three columns: pmcid, pmid, doi
}
\description{
this function is good for retrive data directly
}
\examples{
RetriveMetaDataFromPmidsBatch(c("28852052", "29041955"))

}
