% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveMetaDataFromPmidsBatch}
\alias{RetriveMetaDataFromPmidsBatch}
\title{RetriveMetaDataFromPmidsBatch}
\usage{
RetriveMetaDataFromPmidsBatch(
  pmids,
  apiKey = "",
  email = "",
  outputFileBaseName = "",
  columns = ""
)
}
\arguments{
\item{pmids}{a string of character. PubMed central Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{outputFileBaseName}{a string of characters. The base name of output xml files. If default, there will be no xml saved.}

\item{columns}{the columns of output requested}
}
\value{
a nx7 data frame. With three columns: pmcid, pmid, doi
}
\description{
this function is good for retrive data directly
}
\examples{
RetriveMetaDataFromPmidsBatch(c("28852052", "29041955"))

}
