% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveMetaDataFromPmidsBatch}
\alias{RetriveMetaDataFromPmidsBatch}
\title{RetriveMetaDataFromPmidsBatch}
\usage{
RetriveMetaDataFromPmidsBatch(pmids, apiKey = "",
  endpoint = "esummary", email = "")
}
\arguments{
\item{pmids}{a string of character. PubMed central Id}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{endpoint}{a string of characters. The API endpoint to use. e.g. "esummary", "efetch"}

\item{email}{a string of characters. Your email address}
}
\value{
a nx7 data frame. With three columns: pmcid, pmid, doi
}
\description{
RetriveMetaDataFromPmidsBatch
}
\examples{
RetriveMetaDataFromPmidsBatch(c("28852052", "29041955"), endpoint = "esummary")

}
