% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveMetaDataFromPubmedEfetchParallel}
\alias{RetriveMetaDataFromPubmedEfetchParallel}
\title{RetriveMetaDataFromPubmedEfetchParallel}
\usage{
RetriveMetaDataFromPubmedEfetchParallel(files, columns = "")
}
\arguments{
\item{files}{a list of string. A list of xml filenames}

\item{columns}{the columns of output requested}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
Good for used in local multiple xmls
}
\examples{
baselink <- "https://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi"
link <- paste0(baselink, "?db=pubmed&id=31524133&retmode=xml")
RetriveMetaDataFromPubmedEfetchParallel(link)

}
