% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveMetaDataFromSearch}
\alias{RetriveMetaDataFromSearch}
\title{RetriveMetaDataFromSearch}
\usage{
RetriveMetaDataFromSearch(searchTerm, columns = "", outputFilename = "")
}
\arguments{
\item{searchTerm}{a string of character. search term}

\item{columns}{the columns of output requested}

\item{outputFilename}{a string of characters. Output XML file name}
}
\value{
a list of metaDatarmation retrived from PubMed
}
\description{
RetriveMetaDataFromSearch
}
\examples{
searchTerm <-  "pinkeye"
metaData <- RetriveMetaDataFromSearch(searchTerm)

}
