% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedRead.R
\name{RetriveUrlsFromPmids}
\alias{RetriveUrlsFromPmids}
\title{RetriveUrlsFromPmids}
\usage{
RetriveUrlsFromPmids(pmid, apiKey = "", email = "", fulltext = TRUE)
}
\arguments{
\item{pmid}{a number or a string of characters. The number of pmid.}

\item{apiKey}{a string of characters. The API Key obtained through NCBI account}

\item{email}{a string of characters. Your email address}

\item{fulltext}{a boolean. If TRUE, function only searches for full text link}
}
\value{
a list of characters. A list of urls. Return fulltext urls if fulltext parameter is T
Return NULL if none is found.
}
\description{
Get related url links from given pmid
}
\examples{
RetriveUrlsFromPmids(c("28852052", "29041955"),fulltext = TRUE)

}
