% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{RetriveXmlNodeValuefromDoc}
\alias{RetriveXmlNodeValuefromDoc}
\title{RetriveXmlNodeValuefromDoc}
\usage{
RetriveXmlNodeValuefromDoc(doc, nodePosition, onlyChildren = FALSE)
}
\arguments{
\item{doc}{the parsed XML file}

\item{nodePosition}{the node position of the xml file}

\item{onlyChildren}{whether only use Children of the nodes}
}
\value{
the values of the node
}
\description{
RetriveXmlNodeValuefromDoc
}
\examples{
 
doc <- GetDoc(id = c("5575286", "4804230"),db= "pmc", endpoint="efetch")
RetriveXmlNodeValuefromDoc(doc, "//permissions")

}
